Use pkg-config (when available) to detect -I include paths
Use the PKG-CONFIG-CFLAGS feature of CL-CFFI to automatically detect
the correct include path for freetype. Use :OPTIONAL T to ignore
this when it doesn't work and fall back to the hard-coded paths.

This is a portability fix and in particular it unbreaks cl-freetype2
for the Nix package manager and NixOS Linux distribution.

Changed include patterns from <ftsystem.h> to <freetype/ftsystem.h>
which may cause complications but seems correct (and was necessary to
work in my environment.)